Version 40
This version contains mostly code cleanup in trying to eliminate the modparm.f90 routine.  This is the first revision including those changes.
In this process the "use parm, only : " code was included in each routine using the types or variables that will be removed from modparm.

The following modules were deleted:
  MGTOPS_MODULE.F90
  PLANT_MODULE.F90
  SURF_RUNOFF.MODULE.F90
  EROSION_MODULE.F90
  PESTICIDE_MODULE.F90
  NUTRIENT_MODULE.F90
  EVAPOTRANS_MODULE.F90
  STRUCTURAL_MGT_PRACTICES.F90
  SEPTIC_MODULE.F90
  SOIL_TEMPERATURE_MODULE.F90
  SURFSTORE_MODULE.F90
  WIND_EROSION.F90
  OTHER_MODULE.F90

This routine was deleted because it was not being called or used in the model:
  SUB_SUBBASIN.F90

All of the other routines had the use only (as above) added;

DATA:
LREW_TxtInOut_Landscape was deleted and LREW_TxtInOut_new was copied in its' place;

PLANTS.PLT was updated with plant named TRITICALE added;
HYDROLOGY.HYD - DEP_IMP changed to 5000. in all datasets;
WEATHER-STA.CLI - NUMB column deleted from this file;
ROUTUNIT.RU renamed to ROUTUNIT.RTU
LANDUSE.LUM - CAL_GROUP column added;